Move helper_functions.py
